# api-gateway/api/docs/swagger.yaml
definitions:
...
  genprotos.GetAllPublicVote:
    properties:
      publicVotes:
        items:
          $ref: '#/definitions/genprotos.PublicVote'
        type: array
    type: object
...
  genprotos.PublicVote:
    properties:
      CandidateId:
        type: string
      ElectionId:
        type: string
      Id:
        type: string
      PublicId:
        type: string
    type: object
  genprotos.PublicVoteCreate:
    properties:
      CandidateId:
        type: string
      ElectionId:
        type: string
      PublicId:
        type: string
    type: object
...
  genprotos.Winner:
    properties:
      CandidateId:
        type: string
      ElectionId:
        type: string
      Votes:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /public-vote/{id}:
    get:
      consumes:
      - application/json
      description: Gets a Public Vote by its ID
      operationId: get_by_id_public_votes
      parameters:
      - description: Public Vote ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genprotos.PublicVote'
        "500":
          description: Failed to get public vote by ID
          schema:
            type: string
      summary: Gets a Public Vote by ID
      tags:
      - Public-Votes
  /public-vote/all:
    get:
      consumes:
      - application/json
      description: Gets all Public Votes without any filter
      operationId: get_all_public_votes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genprotos.GetAllPublicVote'
        "500":
          description: Failed to get public votes
          schema:
            type: string
      summary: Gets all Public Votes
      tags:
      - Public-Votes
  /public-vote/create:
    post:
      consumes:
      - application/json
      description: Create Public Vote by reading from body
      operationId: create_public_vote
      parameters:
      - description: Public Vote body data
        in: body
        name: PublicVote
        required: true
        schema:
          $ref: '#/definitions/genprotos.PublicVoteCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Created successfully
          schema:
            type: string
        "500":
          description: Failed to create public vote
          schema:
            type: string
      summary: Creates Public Vote
      tags:
      - Public-Votes
  /public-vote/delete:
    delete:
      consumes:
      - application/json
      description: Deletes a Public Vote by ID
      operationId: delete_public_votes
      parameters:
      - description: Public Vote ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deleted successfully
          schema:
            type: string
        "500":
          description: Failed to delete public vote
          schema:
            type: string
      summary: Deletes a Public Vote
      tags:
      - Public-Votes
  /public-vote/find-winner:
    get:
      consumes:
      - application/json
      description: finds the cnadidate withmost botes
      operationId: fin_the_winner
      parameters:
      - description: election ID
        in: query
        name: election
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genprotos.Winner'
        "500":
          description: Failed to get public vote by ID
          schema:
            type: string
      summary: find the winner
      tags:
      - Public-Votes
  /public-vote/update:
    put:
      consumes:
      - application/json
      description: Updates a Public Vote by ID
      operationId: update_public_votes
      parameters:
      - description: Public Vote ID
        in: query
        name: id
        required: true
        type: string
      - description: Election ID
        in: query
        name: election
        required: true
        type: string
      - description: Public ID
        in: query
        name: public
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Updated successfully
          schema:
            type: string
        "500":
          description: Failed to update public vote
          schema:
            type: string
      summary: Updates a Public Vote
      tags:
      - Public-Votes
...
swagger: "2.0"